+++ NEWS.md
@@ -164,7 +164,7 @@ MKLDNN backend takes advantage of MXNet subgraph to implement the most of possib
 ##### Quantization
 Performance of reduced-precision (INT8) computation is also dramatically improved after the graph optimization feature is applied on CPU Platforms. Various models are supported and can benefit from reduced-precision computation, including symbolic models, Gluon models and even custom models. Users can run most of the pre-trained models with only a few lines of commands and a new quantization script imagenet_gen_qsym_mkldnn.py. The observed accuracy loss is less than 0.5% for popular CNN networks, like ResNet-50, Inception-BN, MobileNet, etc.
 
-Please find detailed information and performance/accuracy numbers here: [MKLDNN README](https://github.com/apache/incubator-mxnet/blob/master/MKLDNN_README.md), [quantization README](https://github.com/apache/incubator-mxnet/tree/master/example/quantization#1) and [design proposal](https://cwiki.apache.org/confluence/display/MXNET/MXNet+Graph+Optimization+and+Quantization+based+on+subgraph+and+MKL-DNN)
+Please find detailed information and performance/accuracy numbers here: [MKLDNN README](https://github.com/apache/incubator-mxnet/blob/master/docs/tutorials/mkldnn/MKLDNN_README.md), [quantization README](https://github.com/apache/incubator-mxnet/tree/master/example/quantization#1) and [design proposal](https://cwiki.apache.org/confluence/display/MXNET/MXNet+Graph+Optimization+and+Quantization+based+on+subgraph+and+MKL-DNN)
 
 ### New Operators
 
@@ -584,7 +584,7 @@ Submodule@commit ID::Last updated by MXNet:: Last update in submodule
 * dlpack@10892ac:: Oct 30, 2017 :: Aug 23, 2018
 * dmlc-core@0a0e8ad:: Aug 15, 2018 :: Nov 15, 2018
 * googletest@ec44c6c:: July 14, 2016 :: July 14, 2016
-* mkldnn@a7c5f53:: Nov 7, 2018 :: Nov 5, 2018
+* mkldnn@722901c:: Feb 13, 2019 :: Feb 12, 2019
 * mshadow@696803b:: Sep 28, 2018 :: Nov 7,  2018
 * onnx-tensorrt@3d8ee04:: Aug 22, 2018 :: Nov 10, 2018
 * openmp@37c7212: Nov 22, 2017 :: Nov 13, 2018